Add Docker support and update project configuration
This commit introduces a Dockerfile for building a Docker image
to streamline the execution of the application. It replaces the old
PipDockerfile, which has been removed, consolidating the build
process into a more efficient flow. Furthermore, it adds `.env`
support in the `.gitignore` to keep sensitive information out of
version control.

The README is updated with instructions for building and running
the Docker container. Several scripts have been added or modified
to assist with the Docker workflow, including building the image,
generating environment variables, and running the application
within the container.

In the Jupyter notebook, minor changes to execution counts and
output formatting ensure compatibility with the new environment,
and dependencies in `requirements.txt` are updated for better
performance and compatibility.