Update README and remove setup script
Updated the README file to correct the S3 URI for storing data.
The previous URI was incorrect and has now been modified to point
to the correct S3 bucket.

Additionally, the setup script has been removed as it is no
longer needed for setting up local dependencies or output
directories. This simplifies the project setup process.